+++ backend/canisters/group/impl/src/lifecycle/post_upgrade.rs
@@ -20,6 +20,9 @@ fn post_upgrade(args: Args) {
     let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
 
+    // Only proceed with removing events from the heap if the stable memory migration is complete
+    assert!(data.stable_memory_event_migration_complete);
+
     if data.chat.events.thread_messages_to_update_in_stable_memory_len() > 0 {
         data.stable_memory_event_migration_complete = false;
     }